backend/requirements.txt
boto3
fastapi
pydantic
python-dotenv
langgraph==0.2.70
uvicorn
SQLAlchemy>=2.0.0
PyMySQL
